feat(spatial types to bathingspot fails):
