+++ __tests__/data/point.json
@@ -1,0 +1,1 @@
+{"type":"Feature","geometry":{"type":"Point","coordinates":[125.6,10.1]},"properties":{"name":"Dinagat Islands"}}
