+++ __tests__/data/polygon.json
@@ -1,1 +1,1 @@
-{"polygon":[[[100,0],[101,0],[101,1],[100,1],[100,0]]]}
+{"type":"Feature","geometry":{"polygon":[[[100,0],[101,0],[101,1],[100,1],[100,0]]]}}
