+++ src/lib/request-handlers/regions/regions-helper.ts
@@ -1,0 +1,27 @@
+import { IObject } from '../../types-interfaces';
+import { isObject } from '../../utils';
+
+const criteria = [
+  {type: 'string', key: 'name'},
+  {type: 'string', key: 'displayName'},
+  {type: 'object', key: 'area'},
+];
+export const createMergeObj: (obj: any) => IObject = (obj) => {
+  const res: IObject = {};
+
+  criteria.forEach((criterion) => {
+    const value = obj[criterion.key];
+    switch (criterion.type) {
+      case 'object':
+      if (isObject(value)) {
+        res.area = value.geometry;
+      }
+      break;
+      default:
+      if (typeof value === criterion.type) {
+        res[criterion.key] = value;
+      }
+    }
+  });
+  return res;
+};
